+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+* v0.0.8
+  - Add `support` dir, to be used for work arounds and arch specific things on dev host
+  - Add `vboxload.sh` to reload network drivers, fixes #13 (current Mavericks + vbox bug)
+
 * v0.0.7
   - Add open source chef server to close #10 (using Ubuntu 12.04)
   - Add chef box